+++ py/fastspecfit/data/emlines.ecsv
