+++ py/fastspecfit/emlines.py
@@ -1,0 +1,705 @@
+"""
+fastspecfit.emlines
+===================
+
+Methods and tools for fitting emission lines.
+
+"""
+import pdb # for debugging
+
+import os, time
+import numpy as np
+import multiprocessing
+
+#from scipy.ndimage import gaussian_filter1d
+#import astropy.units as u
+#from astropy.table import Table, Column, vstack, join, hstack
+from astropy.modeling import Fittable1DModel
+#
+#from fnnls import fnnls
+##from desispec.interpolation import resample_flux
+#from redrock.rebin import trapz_rebin
+
+from scipy import constants
+C_LIGHT = constants.c / 1000.0 # [km/s]
+
+from desiutil.log import get_logger
+log = get_logger()
+
+def read_emlines():
+    """Read the set of emission lines of interest.
+
+    ToDo: add lines to mask during continuum-fitting but which we do not want to
+    emission-line fit.
+
+    """
+    from pkg_resources import resource_filename
+    
+    linefile = resource_filename('fastspecfit', 'data/emlines.ecsv')    
+    linetable = Table.read(linefile, format='ascii.ecsv', guess=False)
+    
+    return linetable    
+
+class EMLineModel(Fittable1DModel):
+    """Class to model the emission-line spectra.
+
+    """
+    from astropy.modeling import Parameter
+
+    # NB! The order of the parameters here matters! linevshift is the shift of
+    # the emission-line velocity (in km/s) with respect to the systemic redshift
+    linevshift_forbidden = Parameter(name='linevshift_forbidden', default=0.0, bounds=(-500.0, 500.0)) # [km/s]
+    linevshift_balmer = Parameter(name='linevshift_balmer', default=0.0, bounds=(-500.0, 500.0)) # [km/s]
+
+    linesigma_forbidden = Parameter(name='linesigma_forbidden', default=50.0, bounds=(5, 350)) # line-sigma [km/s]
+    linesigma_balmer = Parameter(name='linesigma_balmer', default=50.0, bounds=(5, 350)) # line-sigma [km/s]
+
+    # Fragile because the lines are hard-coded--
+    oii_3726_amp = Parameter(name='oii_3726_amp', default=1.0)
+    oii_3729_amp = Parameter(name='oii_3729_amp', default=1.0)
+    oiii_4959_amp = Parameter(name='oiii_4959_amp', default=1.0)
+    oiii_5007_amp = Parameter(name='oiii_5007_amp', default=3.0)
+    nii_6548_amp = Parameter(name='nii_6548_amp', default=1.0)
+    nii_6584_amp = Parameter(name='nii_6584_amp', default=3.0)
+    sii_6716_amp = Parameter(name='sii_6716_amp', default=1.0)
+    sii_6731_amp = Parameter(name='sii_6731_amp', default=1.0)
+    hepsilon_amp = Parameter(name='hepsilon_amp', default=0.5)
+    hdelta_amp = Parameter(name='hdelta_amp', default=0.5)
+    hgamma_amp = Parameter(name='hgamma_amp', default=0.5)
+    hbeta_amp = Parameter(name='hbeta_amp', default=1.0)
+    halpha_amp = Parameter(name='halpha_amp', default=3.0)
+
+    # tie the velocity shifts and line-widths together
+    def tie_vshift(model):
+        return model.linevshift_balmer
+    linevshift_forbidden.tied = tie_vshift
+
+    def tie_sigma(model):
+        return model.linesigma_balmer
+    linesigma_forbidden.tied = tie_sigma
+
+    # tie the [NII] and [OIII] line-strengths together
+    def tie_oiii(model):
+        return model.oiii_5007_amp / 2.8875
+    oiii_4959_amp.tied = tie_oiii
+
+    def tie_nii(model):
+        return model.nii_6584_amp / 2.936
+    nii_6548_amp.tied = tie_nii
+    
+    def __init__(self,
+                 linevshift_forbidden=linevshift_forbidden.default,
+                 linevshift_balmer=linevshift_balmer.default,
+                 linesigma_forbidden=linesigma_forbidden.default,
+                 linesigma_balmer=linesigma_balmer.default,
+                 oii_3726_amp=oii_3726_amp.default, 
+                 oii_3729_amp=oii_3729_amp.default, 
+                 oiii_4959_amp=oiii_4959_amp.default, 
+                 oiii_5007_amp=oiii_5007_amp.default, 
+                 nii_6548_amp=nii_6548_amp.default, 
+                 nii_6584_amp=nii_6584_amp.default, 
+                 sii_6716_amp=sii_6716_amp.default, 
+                 sii_6731_amp=sii_6731_amp.default, 
+                 hepsilon_amp=hepsilon_amp.default, 
+                 hdelta_amp=hdelta_amp.default, 
+                 hgamma_amp=hgamma_amp.default, 
+                 hbeta_amp=hbeta_amp.default, 
+                 halpha_amp=halpha_amp.default,
+                 redshift=None,
+                 emlineR=None, npixpercamera=None,
+                 log10wave=None, **kwargs):
+        """Initialize the emission-line model.
+        
+        emlineR -
+        redshift - required keyword
+        
+        """
+        self.linetable = read_emlines()
+        self.redshift = redshift
+        self.emlineR = emlineR
+        self.npixpercamera = np.hstack([0, npixpercamera])
+
+        # internal wavelength vector for building the emission-line model
+        if log10wave is None:
+            pixkms = 10.0
+            dlogwave = pixkms / C_LIGHT / np.log(10) # pixel size [log-lambda]
+            log10wave = np.arange(np.log10(3000), np.log10(1e4), dlogwave)
+        self.log10wave = log10wave
+            
+        super(EMLineModel, self).__init__(
+            linevshift_forbidden=linevshift_forbidden,
+            linevshift_balmer=linevshift_balmer,
+            linesigma_forbidden=linesigma_forbidden,
+            linesigma_balmer=linesigma_balmer,
+            oii_3726_amp=oii_3726_amp,
+            oii_3729_amp=oii_3729_amp,
+            oiii_4959_amp=oiii_4959_amp,
+            oiii_5007_amp=oiii_5007_amp,
+            nii_6548_amp=nii_6548_amp,
+            nii_6584_amp=nii_6584_amp,
+            sii_6716_amp=sii_6716_amp,
+            sii_6731_amp=sii_6731_amp,
+            hepsilon_amp=hepsilon_amp,
+            hdelta_amp=hdelta_amp,
+            hgamma_amp=hgamma_amp,
+            hbeta_amp=hbeta_amp,
+            halpha_amp=halpha_amp, **kwargs)
+
+    def evaluate(self, emlinewave, *args):
+        """Evaluate the emission-line model.
+        emlineR=None, npixpercamera=None, 
+
+        """ 
+        linevshift_forbidden, linevshift_balmer = args[0], args[1]
+        linez_forbidden = self.redshift + linevshift_forbidden / C_LIGHT
+        linez_balmer = self.redshift + linevshift_balmer / C_LIGHT
+
+        linesigma_forbidden, linesigma_balmer = args[2], args[3]
+        log10sigma_forbidden = linesigma_forbidden / C_LIGHT / np.log(10) # line-width [log-10 Angstrom]
+        log10sigma_balmer = linesigma_balmer / C_LIGHT / np.log(10)       # line-width [log-10 Angstrom]
+
+        lineamps = args[4:]
+        linenames = self.linetable['name'].data
+        isbalmers = self.linetable['isbalmer'].data
+
+        # build the emission-line model [erg/s/cm2/A, observed frame]; should we
+        # multiprocess this step?
+        log10model = np.zeros_like(self.log10wave)
+        for linename, lineamp, isbalmer in zip(linenames, lineamps, isbalmers):
+            restlinewave = self.linetable[self.linetable['name'] == linename]['restwave'][0]
+            if isbalmer:
+                log10sigma = log10sigma_balmer
+                linezwave = np.log10(restlinewave * (1.0 + linez_balmer)) # redshifted wavelength [log-10 Angstrom]
+            else:
+                log10sigma = log10sigma_forbidden
+                linezwave = np.log10(restlinewave * (1.0 + linez_forbidden)) # redshifted wavelength [log-10 Angstrom]
+
+            ww = np.abs(self.log10wave - linezwave) < 20 * log10sigma
+            if np.count_nonzero(ww) > 0:
+                #log.info(linename, 10**linezwave, 10**_emlinewave[ww].min(), 10**_emlinewave[ww].max())
+                log10model[ww] += lineamp * np.exp(-0.5 * (self.log10wave[ww]-linezwave)**2 / log10sigma**2)
+
+        # split into cameras, resample, and convolve with the instrumental
+        # resolution
+        emlinemodel = []
+        for ii in [0, 1, 2]: # iterate over cameras
+            ipix = np.sum(self.npixpercamera[:ii+1])
+            jpix = np.sum(self.npixpercamera[:ii+2])
+            #_emlinemodel = resample_flux(emlinewave[ipix:jpix], 10**self.log10wave, log10model)
+            _emlinemodel = trapz_rebin(10**self.log10wave, log10model, emlinewave[ipix:jpix])
+            
+            if self.emlineR is not None:
+                _emlinemomdel = self.emlineR[ii].dot(_emlinemodel)
+            
+            #plt.plot(10**_emlinewave, _emlinemodel)
+            #plt.plot(10**_emlinewave, self.emlineR[ii].dot(_emlinemodel))
+            #plt.xlim(3870, 3920) ; plt.show()
+            #pdb.set_trace()
+            emlinemodel.append(_emlinemodel)
+            
+        return np.hstack(emlinemodel)
+
+class EMLineFit(object):
+    """Class to fit an emission-line spectrum.
+
+    * https://docs.astropy.org/en/stable/modeling/example-fitting-constraints.html#tied
+    * https://docs.astropy.org/en/stable/modeling/new-model.html
+    * https://docs.astropy.org/en/stable/modeling/compound-models.html#parameters
+
+    """
+    def __init__(self, nball=10, chi2fail=1e6, nproc=1, verbose=False):
+        """Write me.
+        
+        """
+        from astropy.modeling import fitting
+
+        self.verbose = verbose
+        self.nproc = nproc
+        
+        self.nball = nball
+        self.chi2fail = chi2fail
+        self.pixkms = 10.0 # pixel size for internal wavelength array [km/s]
+
+        self.fitter = fitting.LevMarLSQFitter()
+
+    def init_output(self, linetable, nobj=1):
+        """Initialize the output data table for this class.
+
+        """
+        out = Table()
+        out.add_column(Column(name='LINEVSHIFT_FORBIDDEN', length=nobj, dtype='f4'))
+        out.add_column(Column(name='LINEVSHIFT_FORBIDDEN_IVAR', length=nobj, dtype='f4'))
+        out.add_column(Column(name='LINEVSHIFT_BALMER', length=nobj, dtype='f4'))
+        out.add_column(Column(name='LINEVSHIFT_BALMER_IVAR', length=nobj, dtype='f4'))
+        out.add_column(Column(name='LINESIGMA_FORBIDDEN', length=nobj, dtype='f4', unit=u.kilometer / u.second))
+        out.add_column(Column(name='LINESIGMA_FORBIDDEN_IVAR', length=nobj, dtype='f4', unit=u.second**2 / u.kilometer**2))
+        out.add_column(Column(name='LINESIGMA_BALMER', length=nobj, dtype='f4', unit=u.kilometer / u.second))
+        out.add_column(Column(name='LINESIGMA_BALMER_IVAR', length=nobj, dtype='f4', unit=u.second**2 / u.kilometer**2))
+
+        for line in linetable['name']:
+            line = line.upper()
+            out.add_column(Column(name='{}_AMP'.format(line), length=nobj, dtype='f4',
+                                  unit=10**(-17)*u.erg/(u.second*u.cm**2*u.Angstrom)))
+            out.add_column(Column(name='{}_AMP_IVAR'.format(line), length=nobj, dtype='f4',
+                                  unit=10**34*u.second**2*u.cm**4*u.Angstrom**2/u.erg**2))
+            out.add_column(Column(name='{}_FLUX'.format(line), length=nobj, dtype='f4',
+                                  unit=10**(-17)*u.erg/(u.second*u.cm**2)))
+            out.add_column(Column(name='{}_FLUX_IVAR'.format(line), length=nobj, dtype='f4',
+                                  unit=10**34*u.second**2*u.cm**4/u.erg**2))
+            out.add_column(Column(name='{}_BOXFLUX'.format(line), length=nobj, dtype='f4',
+                                  unit=10**(-17)*u.erg/(u.second*u.cm**2)))
+            out.add_column(Column(name='{}_BOXFLUX_IVAR'.format(line), length=nobj, dtype='f4',
+                                  unit=10**34*u.second**2*u.cm**4/u.erg**2))
+            out.add_column(Column(name='{}_CONT'.format(line), length=nobj, dtype='f4',
+                                  unit=10**(-17)*u.erg/(u.second*u.cm**2*u.Angstrom)))
+            out.add_column(Column(name='{}_CONT_IVAR'.format(line), length=nobj, dtype='f4',
+                                  unit=10**34*u.second**2*u.cm**4*u.Angstrom**2/u.erg**2))
+            out.add_column(Column(name='{}_EW'.format(line), length=nobj, dtype='f4',
+                                  unit=u.Angstrom))
+            out.add_column(Column(name='{}_EW_IVAR'.format(line), length=nobj, dtype='f4',
+                                  unit=1/u.Angstrom**2))
+            out.add_column(Column(name='{}_FLUX_LIMIT'.format(line), length=nobj, dtype='f4',
+                                  unit=u.erg/(u.second*u.cm**2)))
+            out.add_column(Column(name='{}_EW_LIMIT'.format(line), length=nobj, dtype='f4',
+                                  unit=u.Angstrom))
+            out.add_column(Column(name='{}_CHI2'.format(line), length=nobj, dtype='f4'))
+            out.add_column(Column(name='{}_NPIX'.format(line), length=nobj, dtype=np.int32))
+
+        return out
+        
+    def chi2(self, bestfit, emlinewave, emlineflux, emlineivar):
+        """Compute the reduced chi^2."""
+        dof = len(emlinewave) - len(bestfit.parameters)
+        emlinemodel = bestfit(emlinewave)
+        chi2 = np.sum(emlineivar * (emlineflux - emlinemodel)**2) / dof
+        return chi2
+
+    def emlinemodel_bestfit(self, specwave, specres, fastspecfit_table):
+        """Wrapper function to get the best-fitting emission-line model
+        from an fastspecfit table (to be used to build QA).
+
+        """
+        npixpercamera = [len(gw) for gw in specwave]
+
+        redshift = fastspecfit_table['Z']
+        linesigma_forbidden = fastspecfit_table['LINESIGMA_FORBIDDEN']
+        linesigma_balmer = fastspecfit_table['LINESIGMA_BALMER']
+
+        linevshift_forbidden = fastspecfit_table['LINEVSHIFT_FORBIDDEN']
+        linevshift_balmer = fastspecfit_table['LINEVSHIFT_BALMER']
+        #linez_forbidden = fastspecfit_table['LINEZ_FORBIDDEN']
+        #linez_balmer = fastspecfit_table['LINEZ_BALMER']
+        #linevshift_forbidden = (linez_forbidden - redshift) * C_LIGHT # [km/s]
+        #linevshift_balmer = (linez_balmer - redshift) * C_LIGHT # [km/s]
+
+        EMLine = EMLineModel(linevshift_forbidden=linevshift_forbidden,
+                             linevshift_balmer=linevshift_balmer,
+                             linesigma_forbidden=linesigma_forbidden,
+                             linesigma_balmer=linesigma_balmer,
+                             redshift=redshift, emlineR=specres,
+                             npixpercamera=npixpercamera)
+        # skip linevshift_[forbidden,balmer] and linesigma_[forbidden,balmer]
+        lineargs = [fastspecfit_table[linename.upper()] for linename in EMLine.param_names[4:]] 
+        lineargs = [linevshift_forbidden, linevshift_balmer, linesigma_forbidden, linesigma_balmer] + lineargs
+
+        _emlinemodel = EMLine.evaluate(np.hstack(specwave), *lineargs)
+
+        # unpack it
+        emlinemodel = []
+        npix = np.hstack([0, npixpercamera])
+        for ii in [0, 1, 2]: # iterate over cameras
+            ipix = np.sum(npix[:ii+1])
+            jpix = np.sum(npix[:ii+2])
+            emlinemodel.append(_emlinemodel[ipix:jpix])
+
+        return emlinemodel
+    
+    def fit(self, data, continuum):
+        """Perform the fit minimization / chi2 minimization.
+        
+        EMLineModel object
+        FC - ContinuumFit object
+
+        need to take into account the instrumental velocity width when computing integrated fluxes
+        
+        """
+        #from scipy import integrate
+        from astropy.stats import sigma_clipped_stats
+        from fastspecfit.continuum import get_d4000
+        
+        # Combine all three cameras; we will unpack them to build the
+        # best-fitting model (per-camera) below.
+        npixpercamera = [len(gw) for gw in data['wave']]
+        npixpercam = np.hstack([0, npixpercamera])
+
+        redshift = data['zredrock']
+        emlinewave = np.hstack(data['wave'])
+        emlineivar = np.hstack(data['ivar'])
+        specflux = np.hstack(data['flux'])
+        emlineflux = specflux - np.hstack(continuum)
+
+        dlogwave = self.pixkms / C_LIGHT / np.log(10) # pixel size [log-lambda]
+        log10wave = np.arange(np.log10(3e3), np.log10(1e4), dlogwave)
+        #log10wave = np.arange(np.log10(emlinewave.min()), np.log10(emlinewave.max()), dlogwave)
+        
+        self.EMLineModel = EMLineModel(redshift=redshift,
+                                       emlineR=data['res'],
+                                       npixpercamera=npixpercamera,
+                                       log10wave=log10wave)
+        nparam = len(self.EMLineModel.parameters)
+
+        #weights = np.ones_like
+        #emlinevar, _ = _ivar2var(emlineivar)
+        #weights[np.logical_not(goodmask)] = 1e16
+        bestfit = self.fitter(self.EMLineModel, emlinewave, emlineflux, weights=np.sqrt(emlineivar))
+
+        emlinemodel = bestfit(emlinewave)
+        chi2 = self.chi2(bestfit, emlinewave, emlineflux, emlineivar).astype('f4')
+
+        # Initialize the output table; see init_fastspecfit for the data model.
+        result = self.init_output(self.EMLineModel.linetable)
+
+        specflux_nolines = specflux - emlinemodel
+
+        d4000_nolines, _ = get_d4000(emlinewave, specflux_nolines, redshift=redshift)
+        result['D4000_NOLINES'] = d4000_nolines
+
+        ## Pack the results in a dictionary and return.
+        ## https://gist.github.com/eteq/1f3f0cec9e4f27536d52cd59054c55f2
+        #result = {
+        #    'converged': False,
+        #    'fit_message': self.fitter.fit_info['message'],
+        #    'nparam': nparam,
+        #    'npix': len(emlinewave),
+        #    'dof': len(emlinewave) - len(self.EMLineModel.parameters),
+        #    'chi2': chi2,
+        #    'linenames': [ll.replace('_amp', '') for ll in self.EMLineModel.param_names[4:]],
+        #}
+        #for param in bestfit.param_names:
+        #    result.update({param: getattr(bestfit, param).value})
+        
+        # uncertainties
+        if self.fitter.fit_info['param_cov'] is not None:
+            cov = self.fitter.fit_info['param_cov']
+            ivar = 1 / np.diag(cov)
+            #result['converged'] = True
+        else:
+            cov = np.zeros((nparam, nparam))
+            ivar = np.zeros(nparam)
+
+        # Need to be careful about uncertainties for tied parameters--
+        # https://github.com/astropy/astropy/issues/7202
+        
+        #err_params = np.sqrt(np.diag(fitter.fit_info['param_cov']))
+        #err = model.copy()
+        #fitting._fitter_to_model_params(err, err_params)            
+            
+        count = 0
+        for ii, pp in enumerate(bestfit.param_names):
+            pinfo = getattr(bestfit, pp)
+            iinfo = getattr(self.EMLineModel, pp)
+
+            # need to think about this more deeply
+            #if pinfo.value == iinfo.value: # not fitted
+            #    result.update({pinfo.name: np.float(0.0)})
+            #else:
+            #    result.update({pinfo.name: pinfo.value.astype('f4')})
+            #result.update({pinfo.name: pinfo.value.astype('f4')})
+            result[pinfo.name.upper()] = pinfo.value.astype('f4')
+                
+            if pinfo.fixed:
+                #result.update({'{}_ivar'.format(pinfo.name): np.float32(0.0)})
+                result['{}_IVAR'.format(pinfo.name.upper())] = pinfo.value.astype('f4')
+            elif pinfo.tied:
+                # hack! see https://github.com/astropy/astropy/issues/7202
+                #result.update({'{}_ivar'.format(pinfo.name): np.float32(0.0)})
+                result['{}_IVAR'.format(pinfo.name.upper())] = np.float32(0.0)
+            else:
+                result['{}_IVAR'.format(pinfo.name.upper())] = ivar[count].astype('f4')
+                #result.update({'{}_ivar'.format(pinfo.name): ivar[count].astype('f4')})
+                count += 1
+
+            #if 'forbidden' in pinfo.name:
+            #    pdb.set_trace()
+
+        # hack for tied parameters---gotta be a better way to do this
+        #result['oiii_4959_amp_ivar'] = result['oiii_5007_amp_ivar'] * 2.8875**2
+        result['OIII_4959_AMP_IVAR'] = result['OIII_5007_AMP_IVAR'] * 2.8875**2
+        result['NII_6548_AMP_IVAR'] = result['NII_6548_AMP_IVAR'] * 2.936**2
+        result['LINEVSHIFT_FORBIDDEN_IVAR'] = result['LINEVSHIFT_BALMER_IVAR']
+        result['LINESIGMA_FORBIDDEN_IVAR'] = result['LINESIGMA_BALMER_IVAR']
+
+        ## convert the vshifts to redshifts
+        #result['linez_forbidden'] = redshift + result['linevshift_forbidden'] / C_LIGHT
+        #result['linez_balmer'] = redshift + result['linevshift_balmer'] / C_LIGHT
+        #result['linez_forbidden_ivar'] = result['linevshift_forbidden_ivar'] * C_LIGHT**2
+        #result['linez_balmer_ivar'] = result['linevshift_balmer_ivar'] * C_LIGHT**2
+
+        # now loop back through and if ivar==0 then set the parameter value to zero
+        if False:
+            if self.fitter.fit_info['param_cov'] is not None:
+                for pp in bestfit.param_names:
+                    if result['{}_IVAR'.format(pp.upper())] == 0.0:
+                        result[pp] = np.float(0.0)
+
+        # get continuum fluxes, EWs, and upper limits
+        sigma_cont = 150.0
+        for oneline in self.EMLineModel.linetable:
+            line = oneline['name'].upper()
+
+            # get the emission-line flux
+            zwave = oneline['restwave'] * (1 + redshift)
+            if oneline['isbalmer']:
+                linesigma = result['LINESIGMA_BALMER']
+            else:
+                linesigma = result['LINESIGMA_FORBIDDEN']
+
+            linesigma_ang = zwave * linesigma / C_LIGHT # [observed-frame Angstrom]
+            norm = np.sqrt(2.0 * np.pi) * linesigma_ang
+
+            result['{}_FLUX'.format(line)] = result['{}_AMP'.format(line)] * norm
+            result['{}_FLUX_IVAR'.format(line)] = result['{}_AMP_IVAR'.format(line)] / norm**2
+
+            # boxcar integration, chi2, and number of pixels
+            lineindx = np.where((emlinewave > (zwave - 3*sigma_cont * zwave / C_LIGHT)) *
+                                (emlinewave < (zwave + 3.*sigma_cont * zwave / C_LIGHT)) *
+                                (emlineivar > 0))[0]
+            npix = len(lineindx)
+            if npix > 0:
+                dof = npix - 3 # ??? [redshift, sigma, and amplitude]
+                chi2 = np.sum(emlineivar[lineindx]*(emlineflux[lineindx]-emlinemodel[lineindx])**2) / dof
+                boxflux = np.sum(emlineflux[lineindx])
+                boxflux_ivar = 1 / np.sum(1 / emlineivar[lineindx])
+            else:
+                npix, chi2, boxflux, boxflux_ivar = 0.0, 1e6, 0.0, 0.0
+
+            result['{}_NPIX'.format(line)] = npix
+            result['{}_CHI2'.format(line)] = chi2
+            result['{}_BOXFLUX'.format(line)] = boxflux
+            result['{}_BOXFLUX_IVAR'.format(line)] = boxflux_ivar
+            
+            # get the continuum and EWs
+            indxlo = np.where((emlinewave > (zwave - 10*sigma_cont * zwave / C_LIGHT)) *
+                              (emlinewave < (zwave - 3.*sigma_cont * zwave / C_LIGHT)))[0]# *
+                              #(emlinemodel == 0))[0]
+            indxhi = np.where((emlinewave < (zwave + 10*sigma_cont * zwave / C_LIGHT)) *
+                              (emlinewave > (zwave + 3.*sigma_cont * zwave / C_LIGHT)))[0]# *
+                              #(emlinemodel == 0))[0]
+            indx = np.hstack((indxlo, indxhi))
+
+            if len(indx) > 5: # require at least 5 pixels to get the continuum level
+                _, cmed, csig = sigma_clipped_stats(specflux_nolines[indx], sigma=3.0)
+                civar = (np.sqrt(len(indx)) / csig)**2
+            else:
+                cmed, civar = 0.0, 0.0
+                
+            result['{}_CONT'.format(line)] = cmed
+            result['{}_CONT_IVAR'.format(line)] = civar
+
+            if result['{}_CONT_IVAR'.format(line)] != 0.0:
+                factor = (1 + redshift) / result['{}_CONT'.format(line)]
+                ew = result['{}_FLUX'.format(line)] * factor   # rest frame [A]
+                ewivar = result['{}_FLUX_IVAR'.format(line)] / factor**2
+
+                # upper limit on the flux is defined by snrcut*cont_err*sqrt(2*pi)*linesigma
+                fluxlimit = np.sqrt(2 * np.pi) * linesigma_ang / np.sqrt(civar)
+                ewlimit = fluxlimit * factor
+            else:
+                ew, ewivar, fluxlimit, ewlimit = 0.0, 0.0, 0.0, 0.0
+
+            result['{}_EW'.format(line)] = ew
+            result['{}_EW_IVAR'.format(line)] = ewivar
+            result['{}_FLUX_LIMIT'.format(line)] = fluxlimit
+            result['{}_EW_LIMIT'.format(line)] = ewlimit
+
+            # simple QA
+            if 'alpha' in line and False:
+                import matplotlib.pyplot as plt
+                _indx = np.where((emlinewave > (zwave - 15*sigma_cont * zwave / C_LIGHT)) *
+                                (emlinewave < (zwave + 15*sigma_cont * zwave / C_LIGHT)))[0]
+                plt.plot(emlinewave[_indx], emlineflux[_indx])
+                plt.plot(emlinewave[_indx], specflux_nolines[_indx])
+                plt.scatter(emlinewave[indx], specflux_nolines[indx], color='red')
+                plt.axhline(y=cmed, color='k')
+                plt.axhline(y=cmed+csig/np.sqrt(len(indx)), color='k', ls='--')
+                plt.axhline(y=cmed-csig/np.sqrt(len(indx)), color='k', ls='--')
+                plt.savefig('junk.png')
+
+        return result
+    
+    def emlineplot(self, data, result, continuum, qadir='.'):
+        """Plot the emission-line spectrum and best-fitting model.
+
+        """
+        from scipy.ndimage import median_filter
+        import matplotlib.pyplot as plt
+        from matplotlib import colors
+        import matplotlib.ticker as ticker
+        import seaborn as sns
+
+        from fastspecfit.util import ivar2var
+
+        redshift = result['Z']
+        _emlinemodel = self.emlinemodel_bestfit(data['wave'], data['res'], result)
+
+        sns.set(context='talk', style='ticks', font_scale=1.3)#, rc=rc)
+
+        col1 = [colors.to_hex(col) for col in ['skyblue', 'darkseagreen', 'tomato']]
+        col2 = [colors.to_hex(col) for col in ['navy', 'forestgreen', 'firebrick']]
+
+        leg = {
+            'targetid': '{} {}'.format(result['TARGETID'], result['FIBER']),
+            'zredrock': '$z_{{\\rm redrock}}$={:.6f}'.format(redshift),
+            'linevshift_forbidden': '$\Delta\,v_{{\\rm forbidden}}$={:.1f} km/s'.format(result['LINEVSHIFT_FORBIDDEN']),
+            'linevshift_balmer': '$\Delta\,v_{{\\rm Balmer}}$={:.1f} km/s'.format(result['LINEVSHIFT_BALMER']),
+            'linesigma_forbidden': '$\sigma_{{\\rm forbidden}}$={:.1f} km/s'.format(result['LINESIGMA_FORBIDDEN']),
+            'linesigma_balmer': '$\sigma_{{\\rm Balmer}}$={:.1f} km/s'.format(result['LINESIGMA_BALMER']),
+            }
+
+        #fig, ax = plt.subplots(1, 4, figsize=(16, 10))#, sharey=True)
+        fig = plt.figure(figsize=(16, 16))
+        gs = fig.add_gridspec(3, 4, height_ratios=[4, 2, 2])
+        #gs = fig.add_gridspec(2, 4, gridspec_kw={'width_ratios': 1.0, 'height_ratios': 0.5})
+
+        # full spectrum
+        bigax = fig.add_subplot(gs[0, :])
+
+        ymin, ymax = 1e6, -1e6
+        for ii in [0, 1, 2]: # iterate over cameras
+            emlinewave = data['wave'][ii]
+            emlineflux = data['flux'][ii] - continuum[ii]
+            emlinemodel = _emlinemodel[ii]
+
+            emlinesigma, good = ivar2var(data['ivar'][ii], sigma=True)
+            emlinewave = emlinewave[good]
+            emlineflux = emlineflux[good]
+            emlinesigma = emlinesigma[good]
+            emlinemodel = emlinemodel[good]
+
+            bigax.fill_between(emlinewave, emlineflux-emlinesigma,
+                               emlineflux+emlinesigma, color=col1[ii], alpha=0.7)
+            bigax.plot(emlinewave, emlinemodel, color=col2[ii], lw=2)
+
+            # get the robust range
+            sigflux, filtflux = np.std(emlineflux), median_filter(emlineflux, 5)
+
+            if -3 * sigflux < ymin:
+                ymin = -2 * sigflux
+            #if np.min(filtflux) < ymin:
+            #    ymin = np.min(filtflux)
+            if np.min(emlinemodel) < ymin:
+                ymin = 0.8 * np.min(emlinemodel)
+                
+            if 5 * sigflux > ymax:
+                ymax = 4 * sigflux
+            if np.max(filtflux) > ymax:
+                ymax = np.max(filtflux)
+            if np.max(emlinemodel) > ymax:
+                ymax = np.max(emlinemodel) * 1.2
+
+        bigax.text(0.95, 0.92, '{}'.format(leg['targetid']), 
+                   ha='right', va='center', transform=bigax.transAxes, fontsize=18)
+        bigax.text(0.95, 0.86, r'{}'.format(leg['zredrock']),
+                   ha='right', va='center', transform=bigax.transAxes, fontsize=18)
+        bigax.text(0.95, 0.80, r'{} {}'.format(leg['linevshift_balmer'], leg['linevshift_forbidden']),
+                   ha='right', va='center', transform=bigax.transAxes, fontsize=18)
+        bigax.text(0.95, 0.74, r'{} {}'.format(leg['linesigma_balmer'], leg['linesigma_forbidden']),
+                   ha='right', va='center', transform=bigax.transAxes, fontsize=18)
+                
+        bigax.set_xlim(3500, 10000)
+        bigax.set_ylim(ymin, ymax)
+        
+        #bigax.set_xlabel(r'Observed-frame Wavelength ($\AA$)') 
+        #bigax.set_ylabel(r'Flux ($10^{-17}~{\rm erg}~{\rm s}^{-1}~{\rm cm}^{-2}~\AA^{-1}$)') 
+        
+        # zoom in on individual emission lines - use linetable!
+        sig = 500.0 # [km/s]
+
+        meanwaves = [np.mean([3730,3727]), 3971, 4103, 4342, 4863, np.mean([4960, 5008]), 6565, np.mean([6718.294, 6732.673])]
+        deltawaves = [0, 0, 0, 0, 0, (5007-4959)/2, (6585-6550)/2, (6733-6718)/2]
+        linenames = [r'[OII]', r'H$\epsilon$', r'H$\delta$', r'H$\gamma$', r'H$\beta$', r'[OIII]', r'H$\alpha$+[NII]', r'[SII]']
+        nline = len(meanwaves)
+
+        removelabels = np.ones(nline, bool)
+        ymin, ymax = np.zeros(nline)+1e6, np.zeros(nline)-1e6
+        
+        ax = []
+        for iax, (meanwave, deltawave, linename) in enumerate(zip(meanwaves, deltawaves, linenames)):
+            if iax < 4:
+                xx = fig.add_subplot(gs[1, iax])
+            else:
+                xx = fig.add_subplot(gs[2, iax-4])
+            ax.append(xx)
+
+            wmin = (meanwave-deltawave)*(1+redshift)-2.5*sig*meanwave/3e5
+            wmax = (meanwave+deltawave)*(1+redshift)+2.5*sig*meanwave/3e5
+
+            # iterate over cameras
+            for ii in [0, 1, 2]: # iterate over cameras
+                emlinewave = data['wave'][ii]
+                emlineflux = data['flux'][ii] - continuum[ii]
+                emlinemodel = _emlinemodel[ii]
+
+                emlinesigma, good = ivar2var(data['ivar'][ii], sigma=True)
+                emlinewave = emlinewave[good]
+                emlineflux = emlineflux[good]
+                emlinesigma = emlinesigma[good]
+                emlinemodel = emlinemodel[good]
+
+                indx = np.where((emlinewave > wmin) * (emlinewave < wmax))[0]
+                #log.info(ii, linename, len(indx))
+                if len(indx) > 1:
+                    removelabels[iax] = False
+                    xx.fill_between(emlinewave[indx], emlineflux[indx]-emlinesigma[indx],
+                                    emlineflux[indx]+emlinesigma[indx], color=col1[ii], alpha=0.5)
+                    xx.plot(emlinewave[indx], emlinemodel[indx], color=col2[ii], lw=3)
+
+                    # get the robust range
+                    sigflux, filtflux = np.std(emlineflux[indx]), median_filter(emlineflux[indx], 3)
+
+                    _ymin, _ymax = -1.5 * sigflux, 3 * sigflux
+                    if np.max(emlinemodel[indx]) > _ymax:
+                        _ymax = np.max(emlinemodel[indx]) * 1.2
+                    if np.max(filtflux) > _ymax:
+                        _ymax = np.max(filtflux)
+                    if np.min(emlinemodel[indx]) < _ymin:
+                        _ymin = 0.8 * np.min(emlinemodel[indx])
+                        
+                    if _ymax > ymax[iax]:
+                        ymax[iax] = _ymax
+                    if _ymin < ymin[iax]:
+                        ymin[iax] = _ymin
+                    
+                xx.text(0.08, 0.9, linename, ha='left', va='center',
+                        transform=xx.transAxes, fontsize=20)
+                    
+        for iax, xx in enumerate(ax):
+            if removelabels[iax]:
+                xx.set_ylim(0, 1)
+                xx.set_xticklabels([])
+                xx.set_yticklabels([])
+            else:
+                xx.set_ylim(ymin[iax], ymax[iax])
+                xlim = xx.get_xlim()
+                #log.info(linenames[iax], xlim, np.diff(xlim))
+                xx.xaxis.set_major_locator(ticker.MaxNLocator(3))
+                #xx.xaxis.set_major_locator(ticker.MultipleLocator(20)) # wavelength spacing of ticks [Angstrom]
+                #if iax == 2:
+                #    pdb.set_trace()
+
+        # common axis labels
+        tp, bt, lf, rt = 0.95, 0.09, 0.12, 0.95
+        
+        fig.text(lf-0.07, (tp-bt)/2+bt,
+                 r'Flux ($10^{-17}~{\rm erg}~{\rm s}^{-1}~{\rm cm}^{-2}~\AA^{-1}$)',
+                 ha='center', va='center', rotation='vertical')
+        fig.text((rt-lf)/2+lf, bt-0.06, r'Observed-frame Wavelength ($\AA$)',
+                 ha='center', va='center')
+            
+        plt.subplots_adjust(wspace=0.27, top=tp, bottom=bt, left=lf, right=rt, hspace=0.22)
+
+        pngfile = os.path.join(qadir, 'emlinefit-{}-{}-{}.png'.format(
+            result['TILE'], result['NIGHT'], result['TARGETID']))
+        log.info('Writing {}'.format(pngfile))
+        fig.savefig(pngfile)
+        plt.close()
