+++ py/fastspecfit/util.py
@@ -1,0 +1,21 @@
+"""
+fastspecfit.util
+================
+
+General utilities.
+
+"""
+import numpy as np
+
+def ivar2var(ivar, sigma=False):
+    """Safely convert an inverse variance to a variance."""
+    var = np.zeros_like(ivar)
+    goodmask = ivar > 0 # True is good
+    if np.count_nonzero(goodmask) == 0:
+        log.warning('All values are masked!')
+        raise ValueError
+    var[goodmask] = 1 / ivar[goodmask]
+    if sigma:
+        var = np.sqrt(var) # return a sigma
+    return var, goodmask
+
